[up:conv-dilation] move cached_property into a separate file
