+++ nujo/autodiff/_functions/_transform.py
@@ -3,6 +3,7 @@ from typing import List, Optional, Tuple, Union
 
 from numpy import add, arange, ndarray, pad, repeat, tile, zeros
 
+from nujo._cache import cached_property
 from nujo.autodiff.function import Function
 from nujo.autodiff.tensor import Tensor
 
@@ -12,32 +13,6 @@ __all__ = [
     '_ConstPad',
     '_Im2col',
 ]
-
-
-class cached_property(object):
-    ''' A property that is only computed once per instance and then replaces itself
-    with an ordinary attribute.
-
-    Deleting the attribute resets the property.
-
-    '''
-    def __init__(self, func):
-        self.__doc__ = getattr(func, '__doc__')
-        self.func = func
-
-    def __get__(self, obj, cls):
-        if obj is None:
-            return self
-
-        key = self.func.__name__
-        cache = obj.__dict__
-
-        if key in cache:
-            return cache[key]
-        else:
-            value = cache[key] = self.func(obj)
-            return value
-
 
 # ====================================================================================================
